Add util for dateTime/pad
